urgent commit

+++ app/src/main/java/com/example/apolusov/kotlintest/PointD.kt
@@ -1,8 +1,8 @@
 package com.example.apolusov.kotlintest
 
 data class PointD(
-    val x: Int,
-    val y: Int,
+    val x: Float,
+    val y: Float,
     val text: Int
 ) {
     override fun toString() = "[$x, $y, $text]"
